+++ lib/core/utils/custom_widgets/custom_text_widget.dart
@@ -1,15 +1,6 @@
 // custom for text and textstyle
 import 'package:flutter/material.dart';
 
-_fontSizeForEveryDevice(BuildContext context) {
-  final orientations = MediaQuery.of(context).orientation;
-  final isLandscape = orientations == Orientation.landscape;
-  bool _isTablet(BuildContext context) {
-    final shortestSide = MediaQuery.of(context).size.shortestSide;
-    return shortestSide >= 600;
-  }
-}
-
 Widget customText({
   String? text,
   double? fontSize,
